# backend/internal/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.createProjectRequest:
    properties:
      default_environment:
        description: DefaultEnvironment names the first environment; defaults to "prod".
        type: string
      display_name:
        description: DisplayName is the human label; defaults to Slug when empty.
        type: string
      org_id:
        description: |-
          OrgID is the owning org. Empty → the caller's personal org (their username).
          A non-empty value requires the caller to be Owner/Admin of that org.
        type: string
      slug:
        description: Slug is the project name (DNS-1123-label-safe). Unique platform-wide.
        type: string
    required:
    - slug
    type: object
...
info:
  contact: {}
  description: |-
    Platform API to order and manage cloud resources — managed databases, apps, VMs (app servers), AI models, and public API endpoints — on the DADA Cloud platform.
    Mutations are asynchronous: create/update/delete enqueue an operation and return 202 Accepted with an operation object; poll GET /projects/{projectId}/operations/{operationId} until the operation reaches a terminal status (Ready or Failed).
  title: DADA Cloud Console API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: Returns every project the authenticated user is a member of, each
        annotated with the caller's role in that project. Read-only. Start here to
        discover project IDs for other calls.
      operationId: listProjects
      produces:
      - application/json
      responses:
        "200":
          description: object with a projects array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects the caller can access
      tags:
      - project
    post:
      consumes:
      - application/json
      description: Creates a project plus its default environment. Without org_id
        the project goes into your personal org (you are Owner). With org_id you must
        be Owner/Admin of that org. The slug must be DNS-1123-label-safe and is unique
        platform-wide.
      operationId: createProject
      parameters:
      - description: Project to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: object with project_id, default_environment_id, org_id and
            role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: slug already taken
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - project
  /projects/{projectId}:
    delete:
      description: 'Destructive, asynchronous: enqueues a DeleteProject operation
        that removes the project''s entire git tree (every environment/app/resource)
        in one commit and wipes the project''s DB rows. Call the project delete-impact
        endpoint first to preview the blast radius. Returns 202 with an operation;
        poll it until terminal.'
      operationId: deleteProject
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: object with the accepted operation
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      description: Returns one project, the caller's role, and the project's environments
        (each with id, name, namespace, runtime). Read-only. Use the returned environment
        IDs for app/database/model calls.
      operationId: getProject
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with project, role and environments
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a project with its environments
      tags:
      - project
...
  /projects/{projectId}/environments/{envId}/namespace-policy:
    put:
      consumes:
      - application/json
      description: 'Updates the Kubernetes LimitRange and ResourceQuota for an environment''s
        namespace. Admin-only (platform-admin or client-admin). Asynchronous: returns
        202 with an operation id; poll the operation endpoint until terminal.'
      operationId: setNamespacePolicy
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Environment UUID
        in: path
        name: envId
        required: true
        type: string
      - description: Object with limit_range and resource_quota JSON specs
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: object with the accepted operation id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set an environment's namespace LimitRange + ResourceQuota
      tags:
      - project
...
  /projects/{projectId}/operations:
    get:
      description: Returns the 50 most recent async operations for a project (newest
        first), with their current status. Read-only. Use this to track the outcome
        of create/update/delete calls.
      operationId: listOperations
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object with an operations array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List recent operations in a project
      tags:
      - operation
...
  /projects/default:
    post:
      description: Returns the caller's default project, provisioning one (in your
        personal org, you are Owner) when you have zero projects. Idempotent — repeated
        calls return the same project.
      operationId: ensureDefaultProject
      produces:
      - application/json
      responses:
        "200":
          description: object with project_id, default_environment_id, org_id and
            role
          schema:
            additionalProperties: true
            type: object
        "201":
          description: newly created default project
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get or create the caller's default project
      tags:
      - project
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"